+++ propagator/opentelemetry-propagator-b3/src/opentelemetry/propagators/b3/__init__.py
@@ -17,7 +17,7 @@ from re import compile as re_compile
 
 import opentelemetry.trace as trace
 from opentelemetry.context import Context
-from opentelemetry.trace.propagation.textmap import (
+from opentelemetry.propagators.textmap import (
     Getter,
     Setter,
     TextMapPropagator,