+++ CHANGELOG.md
@@ -24,6 +24,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1661])(https://github.com/open-telemetry/opentelemetry-python/pull/1661)
 - Fixed distro configuration with `OTEL_TRACES_EXPORTER` env var set to `otlp`
   ([#1657])(https://github.com/open-telemetry/opentelemetry-python/pull/1657)
+- Moving `Getter`, `Setter` and `TextMapPropagator` out of `opentelemetry.trace.propagation` and
+  into `opentelemetry.propagators`
+  ([#1662])(https://github.com/open-telemetry/opentelemetry-python/pull/1662)
+- Rename `BaggagePropagator` to `W3CBaggagePropagator`
+  ([#1663])(https://github.com/open-telemetry/opentelemetry-python/pull/1663)
 
 ## [0.18b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v0.18b0) - 2021-02-16